requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
jinja2==3.1.2
python-dotenv==1.0.0
boto3==1.28.64
pydantic==2.4.2
